test(refactor): helper method 분리
